fix: Prevent whole minicard from dissapearing after clicking star and delete specified minicard

+++ main.js
@@ -12,12 +12,22 @@ var cardGrid = document.querySelector(".card-grid");
 saveButton.addEventListener("click", function(e) {
   makeNewIdea(e);
 });
-cardGrid.addEventListener("click", function(e) {
-  deleteMiniCard(e);
+// cardGrid.addEventListener("click", function(e) {
+//   deleteMiniCard(e);
+// });
+// cardGrid.addEventListener('click', function(e) {
+//   makeFavorite(e)
+// });
+cardGrid.addEventListener('click', function (e) {
+  if (e.target.classList.contains('white-star')) {
+    makeFavorite(e);
+  } else if (e.target.classList.contains('orange-star')) {
+      UnFavorite(e);
+  } else if (e.target.classList.contains('delete-button')) {
+    deleteMiniCard(e);
+  }
 });
-cardGrid.addEventListener('click', function(e) {
-  makeFavorite(e)
-});
+
 
 saveButton.addEventListener('mouseover', function(){
     if(!saveButton.disabled){
@@ -76,16 +86,20 @@ function checkInput(){
         saveButton.disabled = true;
     }
 }
-
+// var whiteStar 
+var orangeStar 
 
 function showCards() {
   cardGrid.innerHTML = "";
   for (var i = 0; i < ideas.length; i++) {
+ 
+
+    
     cardGrid.innerHTML += `
         <article class="mini-card" id="${ideas[i].id}">
             <header class="mini-card-header">
-              <button class="orange-star hidden"></button> 
-              <button class="white-star"></button> 
+              <button class="orange-star hidden "></button> 
+              <button class="white-star "></button> 
               <button class="delete-button"></button>
             </header>
             <h2> ${ideas[i].title} </h2>
@@ -97,38 +111,64 @@ function showCards() {
 // var whiteStar = !ideas[i].isFavorite;
 // var orangeStar = ideas[i].isFavorite;
 {/* <button class="orange-star hidden ${orangeStar}"></button> 
-<button class="white-star ${whiteStar}"></button> */}
+<button class="white-star ${orangeStar}""></button> */}
 
 function deleteMiniCard(e) {
   for (var i = 0; i < ideas.length; i++) {
-    if (e.target.classList.contains('delete-button')) {
+    if (parseInt(e.target.parentElement.parentElement.id) === ideas[i].id) {
+      console.log(e.target)
       ideas.splice(i, 1);
       showCards()
-      return
+      //return
     }
   }
+   
 }
+
+// parseInt(e.target.closest('article').id)
+// e.target.classList.contains('delete-button')
 
 function toggleClass(element, className) {
   element.classList.toggle(className)
 }
 
+// function makeFavorite(e) {
+//   for (var i = 0; i < ideas.length; i++) {
+//     if (e.target.classList.contains('white-star')) {
+//       console.log(e.target.classList.contains('white-star'))
+//       ideas[i].isFavorite = true
+//       toggleClass(e.target, 'hidden');
+//       toggleClass(e.target.parentElement.firstElementChild, 'hidden');
+//     }
+//     if (e.target.classList.contains('orange-star')) {
+//       console.log('potato')
+//       ideas[i].isFavorite = false;
+//       toggleClass(e.target, 'hidden');
+//       toggleClass(e.target.parentElement.children[1], 'hidden');
+   
+//     }
+    
+//    return
+//   } 
+// }
+
 function makeFavorite(e) {
   for (var i = 0; i < ideas.length; i++) {
-    if (e.target.classList.contains('white-star')) {
+    if (parseInt(e.target.closest('.mini-card').id) === ideas[i].id) {
       ideas[i].isFavorite = true;
       toggleClass(e.target, 'hidden');
-      toggleClass(e.target.parentElement.firstElementChild, 'hidden')
+      toggleClass(e.target.parentElement.firstElementChild, 'hidden');
     }
-    if (e.target.classList.contains('orange-star')) {
-      console.log('potato')
-      ideas[i].isFavorite = false;
-      toggleClass(e.target, 'hidden');
-      toggleClass(e.target.parentElement.children[1], 'hidden')
-    }
-    return
-  } 
+  }
 }
 
-
-
+function UnFavorite(e) {
+  for (var i = 0; i < ideas.length; i++) {
+    if (parseInt(e.target.closest('.mini-card').id) === ideas[i].id) {
+      console.log(e.target)
+      ideas[i].isFavorite = false;
+      toggleClass(e.target, 'hidden');
+      toggleClass(e.target.parentElement.children[1], 'hidden');
+    }
+  }
+}
